Moved local state from BookStore component to individual components
