+++ src/components/Pages/BookStore/Budget.jsx
@@ -1,4 +1,8 @@
-const Budget = ({ budget }) => {
+import React, { useState } from "react";
+
+const Budget = () => {
+  const [budget, setBudget] = useState(100);
+
   return <h2>Your budget is: {budget}$</h2>;
 };
